ㅅㅅ

+++ src/main/java/io/hhplus/tdd/domain/order/presentation/dto/req/MakeOrderReqDTO.java
@@ -13,8 +13,7 @@ public record MakeOrderReqDTO(
     public record ItemInfo(
             Long productId,
             Long productOptionId,
-            Long quantity
+            int quantity
     ){
-
     }
 }
